Add ntype.py, LocalRom now inherits from BigStream
