+++ Rom.py
@@ -7,20 +7,20 @@ import struct
 import subprocess
 import random
 import copy
+from ntype import BigStream
 
 from Utils import local_path, default_output_path
 
-DMADATA_START = 0x0
+DMADATA_START = 0x1A500
 
-class LocalRom(object):
+class LocalRom(BigStream):
     def __init__(self, settings, patch=True):
-        self.last_address = None
+        super().__init__([])
 
         file = settings.rom
         decomp_file = 'MMDEC.z64'
 
         os.chdir(os.path.dirname(os.path.realpath(__file__)))
-        #os.chdir(output_path(os.path.dirname(os.path.realpath(__file__))))
         
         with open(local_path('data/generated/symbols.json'), 'r') as stream:
             symbols = json.load(stream)
@@ -80,81 +80,6 @@ class LocalRom(object):
 
     def sym(self, symbol_name):
         return self.symbols.get(symbol_name)
-
-    def seek_address(self, address):
-        self.last_address = address
-
-    def read_byte(self, address):
-        self.last_address = address + 1
-        return self.buffer[address]
-
-    def read_bytes(self, address, len):
-        self.last_address = address + len
-        return self.buffer[address : address + len]
-
-    def read_int16(self, address):
-        return bytes_as_int16(self.read_bytes(address, 2))
-
-    def read_int24(self, address):
-        return bytes_as_int24(self.read_bytes(address, 3))
-
-    def read_int32(self, address):
-        return bytes_as_int32(self.read_bytes(address, 4))
-
-    def write_byte(self, address, value):
-        if address == None:
-            address = self.last_address
-        self.buffer[address] = value
-        self.last_address = address + 1
-
-    def write_sbyte(self, address, value):
-        if address == None:
-            address = self.last_address
-        self.write_bytes(address, struct.pack('b', value))
-
-    def write_int16(self, address, value):
-        if address == None:
-            address = self.last_address
-        self.write_bytes(address, int16_as_bytes(value))
-
-    def write_int24(self, address, value):
-        if address == None:
-            address = self.last_address
-        self.write_bytes(address, int24_as_bytes(value))
-
-    def write_int32(self, address, value):
-        if address == None:
-            address = self.last_address
-        self.write_bytes(address, int32_as_bytes(value))
-
-    def write_f32(self, address, value:float):
-        if address == None:
-            address = self.last_address
-        self.write_bytes(address, struct.pack('>f', value))
-
-    def write_bytes(self, startaddress, values):
-        if startaddress == None:
-            startaddress = self.last_address
-        for i, value in enumerate(values):
-            self.write_byte(startaddress + i, value)
-
-    def write_int16s(self, startaddress, values):
-        if startaddress == None:
-            startaddress = self.last_address
-        for i, value in enumerate(values):
-            self.write_int16(startaddress + (i * 2), value)
-
-    def write_int24s(self, startaddress, values):
-        if startaddress == None:
-            startaddress = self.last_address
-        for i, value in enumerate(values):
-            self.write_int24(startaddress + (i * 3), value)
-
-    def write_int32s(self, startaddress, values):
-        if startaddress == None:
-            startaddress = self.last_address
-        for i, value in enumerate(values):
-            self.write_int32(startaddress + (i * 4), value)
 
     def write_to_file(self, file):
         # self.verify_dmadata()
@@ -257,25 +182,3 @@ class LocalRom(object):
 
         else:
             rom.write_int32s(cur, [start, end, start, 0])
-
-
-def int16_as_bytes(value):
-    value = value & 0xFFFF
-    return [(value >> 8) & 0xFF, value & 0xFF]
-
-def int24_as_bytes(value):
-    value = value & 0xFFFFFF
-    return [(value >> 16) & 0xFF, (value >> 8) & 0xFF, value & 0xFF]
-
-def int32_as_bytes(value):
-    value = value & 0xFFFFFFFF
-    return [(value >> 24) & 0xFF, (value >> 16) & 0xFF, (value >> 8) & 0xFF, value & 0xFF]
-
-def bytes_as_int16(values):
-    return (values[0] << 8) | values[1]
-
-def bytes_as_int24(values):
-    return (values[0] << 16) | (values[1] << 8) | values[2]
-
-def bytes_as_int32(values):
-    return (values[0] << 24) | (values[1] << 16) | (values[2] << 8) | values[3]
